% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_abund.R
\name{ag.abund}
\alias{ag.abund}
\title{Aggregate abundance samples}
\usage{
ag.abund(x, ag.var)
}
\arguments{
\item{x}{A sample abundance data frame produced by \code{\link{sample.abund}}}

\item{ag.var}{A factor (or character) variable in \code{x} that indicates the group membership
for aggregation.}
}
\description{
Takes abundance samples created with the function
\code{\link{sample.abund}} and aggregates them based on a factor variable in
(or added) to the \code{sample.abund} data set.
}
\author{
Devin S. Johnson
}
