% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_summary.R
\name{ag.summary}
\alias{ag.summary}
\title{Summarize count posterior distributions}
\usage{
ag.summary(x, ci.prob = 0.9)
}
\arguments{
\item{x}{A data frame created with the sampling function \code{\link{sample.abund}} or
\code{\link{ag.abund}}}

\item{ci.prob}{Probability level for credible intervals. Defaults to 0.9.}
}
\description{
The posterior distribution of the count sample is summarized for
each site. The resulting data frame can be used for site-level plots, etc.
}
\author{
Devin S. Johnson
}
