% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_trend.R
\name{ag.trend}
\alias{ag.trend}
\title{Estimate trends of aggregated abundance}
\usage{
ag.trend(x, timeframe, ci.prob = 0.9)
}
\arguments{
\item{x}{A abundance sample data frame from \code{\link{sample.abund}} or \code{\link{ag.abund}}}

\item{timeframe}{a 2-vector giving the start year and end year of the desired
trend estimate, e.g., \code{timeframe = c(1989,2019)}.}

\item{ci.prob}{The probability for the credible interval. Defaults to 0.9.}
}
\description{
The growth trend of is estimated from the samples obtained from the
functions \code{\link{sample.abund}} and \code{\link{ag.abund}}.
}
\details{
The function returns a named list with elements \code{growth}, \code{fitted},
and \code{sample}. The \code{growth} element contains a table with the estimated
growth of each aggregation in \emph{percent growth} form. The \code{fitted} element contains a
table with the fitted trendline values on the count scale. This can be used for plotting.
Finally, the \code{sample} element contains the posterior sample. This can be used for further
analysis of trends such as comparisons, etc.
}
\author{
Devin S. Johnson
}
