% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gam.R
\name{fit.gam}
\alias{fit.gam}
\title{Fit penalized GAM to each site for missing data imputation}
\usage{
fit.gam(data, obl.corr, alt.mod = NULL, warn = FALSE, debug = FALSE, ...)
}
\arguments{
\item{data}{A data frame containing the sites, counts, and years.}

\item{obl.corr}{Logical. Should oblique photo correction be used?}

\item{alt.mod}{An alternative to the defaul model. See Details.}

\item{warn}{Logical. Should fitting warnings be printed. Defaults to \code{FALSE}.}

\item{debug}{Logical. If set to \code{TRUE} the function will drop into browser
mode within the function upon execution. Defaults to \code{FALSE}.}

\item{...}{Extra arguments passed to \code{\link[mgcv]{gam}}}
}
\description{
This function takes a processed data frame with counts during survey
years for each site and fits a multi-site GAM using a Tweedie response distribution.
}
\details{
A penalized GAM model is fit to the data with a Tweedie distribution for the response distribution.
The mean function for site i in year t is modeled with a smooth year term for all sites plus
a factor smooth such that each site has its own smooth year function as well the \code{mgcv} formula is
\code{mu.form = count ~ s(year, SITE, bs='fs', k=8, m=1)}. For the power and dispersion formulas,
\code{p.form = ~ s(SITE, bs='re')} and \code{phi.form = ~ s(SITE, bs='re')}, so that each
site has its own p and phi parameters for the Tweedie specification. Thus, for each site, the
mean count is mu(i,t) and the variance is V(i,t) = phi(i)*mu(i,t)^p(i). To specify a different formula for the model, set
\code{alt.model = list(mu=mu.form.alt, p=p.form.alt, phi=phi.form.alt)} where the \code{*.alt} signifies
the formula for the desired alternative.

If \code{obl.corr = TRUE} then a model will be fitted that corrects for the approximately
3.8\% reduction in the expected count if the data were collected using an oblique
photo vs. a medium format vertical photo. Unlike the original \code{agTrend} package
the uncertainty in this estimate is not accounted for in the model. A few initial tests
revealed that is source of variation seems insignificant when compared to the
natural variation of the observed counts and the model was significantly easier to
fit and more robust when fixing this quantity.
}
\author{
Devin S. Johnson
}
