% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_data.R
\name{proc.data}
\alias{proc.data}
\title{Process SSL count data .xlsx file}
\usage{
proc.data(allcounts, age, dps)
}
\arguments{
\item{allcounts}{the file path of the current allcounts excel file.}

\item{age}{Age needed, must be one of 'pup' or 'nonpup'}

\item{dps}{DPS needed, must be one of 'wdps' or 'edps'}
}
\description{
The current 'ALLCOUNTS_v(x).xlsx' file is processed into the
separate, R friendly, data sets necessary for trend analysis.
}
\author{
Devin S. Johnson
}
