% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_abund.R
\name{sample.abund}
\alias{sample.abund}
\title{Sample missing counts}
\usage{
sample.abund(
  fit,
  data,
  yrs,
  size = 1000,
  add.site.data = NULL,
  keep.par = FALSE,
  debug = FALSE
)
}
\arguments{
\item{fit}{A fitted model object from a call to \code{\link{fit.gam}}.}

\item{data}{Data used for model fitting.}

\item{yrs}{Years used for simulation. Defaults to a sequence of years from the
first to the last. But this can be any subset of years in the survey window.}

\item{size}{The sample size of the random imputation draws. Defaults to 1,000.}

\item{add.site.data}{An additional data set with a column lableled 'site' matching
the sites used in the model fitting data that contain additional groupings of other
site-level data.}

\item{keep.par}{Logical. Should the sample of parameters used be retained? It is not often needed
for further analysis.}

\item{debug}{Enter into the function for debugging.}
}
\description{
Using a fitted GAM SSL model (see \code{\link{fit.gam}}) and the data used
for fitting, missing survey counts are imputed for aggregated trend and abundance
analysis.
}
\author{
Devin S. Johnson
}
